requirements.txt
# Core
python-dotenv
pydantic
pydantic-settings

# Logging
loguru

# Web
httpx
beautifulsoup4
playwright

# Database (MVP will use SQLite, but good to have base SQLAlchemy)
sqlalchemy
aiosqlite # For async SQLite

# CLI with Rich UI
typer[all]
rich

# Testing
pytest
pytest-mock
pytest-cov
pytest-html # For HTML test reports
pytest-asyncio # For async tests

# Utilities
tldextract
respx

# Phone number parsing
phonenumbers

# Email validation (optional but recommended)
email-validator

# Address parsing (optional)
# usaddress  # US address parsing - uncomment if needed
# spacy # NLP for entity extraction - uncomment if needed

# Note: For spaCy, also run: python -m spacy download en_core_web_sm

# LLM Processing (NEW)
llama-cpp-python>=0.2.0  # For Qwen2 Instruct 7B
markdownify>=0.11.0      # HTML to Markdown conversion for LLM efficiency
huggingface-hub>=0.19.0  # For model downloading (optional)

# Enhanced HTML Processing
html2text>=2020.1.16    # Alternative HTML to text conversion
lxml>=4.9.0             # Faster HTML parsing

# Production optimizations
uvloop>=0.19.0          # Faster event loop for Linux/Mac
orjson>=3.9.0           # Faster JSON processing
